batman doc
git-svn-id: svn://scm.r-forge.r-project.org/svnroot/batman@233 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ pkg/batman/R/batman.R
@@ -277,9 +277,14 @@ batman<-function(BrukerDataDir, BrukerDataZipDir, txtFile, rData, createDir = TR
     plotRelCon(BM, saveFigDir = dirctime)
   if (figMetaFit)
     plotMetaFit(BM, saveFigDir = dirctime)
-  
-  file.remove(dir3)
-  file.remove(dir4)
+  if (file.exists(dir3))
+  {
+    file.remove(dir3)
+  }
+  if (file.exists(dir4))
+  {
+    file.remove(dir4)
+  }
   if (chemshif == 1)
   {
     if (file.exists(dir6))
